move find_related_easyconfigs into framework.easyconfigs.tools + minor enhancements
