+++ easybuild/framework/easyconfig/tools.py
@@ -36,10 +36,12 @@ alongside the EasyConfig class to represent parsed easyconfig files.
 @author: Fotis Georgatos (Uni.Lu, NTUA)
 @author: Ward Poelmans (Ghent University)
 """
-
+import glob
 import os
+import re
 import sys
 import tempfile
+from distutils.version import LooseVersion
 from vsc.utils import fancylogger
 
 # optional Python packages, these might be missing
@@ -68,7 +70,7 @@ except ImportError, err:
     graph_errors.append("Failed to import graphviz: try yum install graphviz-python, or apt-get install python-pygraphviz")
 
 from easybuild.framework.easyconfig import EASYCONFIGS_PKG_SUBDIR
-from easybuild.framework.easyconfig.easyconfig import ActiveMNS, find_related_easyconfigs, process_easyconfig
+from easybuild.framework.easyconfig.easyconfig import ActiveMNS, create_paths, process_easyconfig
 from easybuild.tools.build_log import EasyBuildError
 from easybuild.tools.config import build_option
 from easybuild.tools.filetools import find_easyconfigs, write_file
@@ -77,6 +79,7 @@ from easybuild.tools.modules import modules_tool
 from easybuild.tools.multi_diff import multi_diff
 from easybuild.tools.ordereddict import OrderedDict
 from easybuild.tools.run import run_cmd
+from easybuild.tools.toolchain import DUMMY_TOOLCHAIN_NAME
 from easybuild.tools.utilities import quote_str
 
 
@@ -347,6 +350,78 @@ def stats_to_str(stats):
     return txt
 
 
+def find_related_easyconfigs(path, ec):
+    """
+    Find related easyconfigs for provided parsed easyconfig in specified path.
+
+    A list of easyconfigs for the same software (name) is returned,
+    matching the 1st criterion that yields a non-empty list.
+
+    Software version is considered more important than toolchain.
+
+    Toolchain is considered with exact same version prior to without version (only name).
+
+    Matching versionsuffix is considered prior to any versionsuffix.
+
+    Exact software versions are considered prior to matching major/minor version numbers,
+    and only matching major version number, before any software version is considered.
+
+    The following criteria are considered, in order (with 'version criterion' being either an
+    exact version match, a major/minor version match, a major version match, or no version match).
+
+    (i)   software version criterion, versionsuffix and toolchain name/version
+    (ii)  software version criterion, versionsuffix and toolchain name (any toolchain version)
+    (iii) software version criterion, versionsuffix (any toolchain name/version)
+    (iv)  software version criterion and toolchain name/version (any versionsuffix)
+    (v)   software version criterion and toolchain name (any versionsuffix, toolchain version)
+    (vi)  software version criterion (any versionsuffix, toolchain name/version)
+
+    If no related easyconfigs with a matching software name are found, an empty list is returned.
+    """
+    name = ec.name
+    version = ec.version
+    versionsuffix = ec['versionsuffix']
+    toolchain_name = ec['toolchain']['name']
+    toolchain_name_pattern = r'-%s-\S+' % toolchain_name
+    toolchain = "%s-%s" % (toolchain_name, ec['toolchain']['version'])
+    toolchain_pattern = '-%s' % toolchain
+    if toolchain_name == DUMMY_TOOLCHAIN_NAME:
+        toolchain_name_pattern = ''
+        toolchain_pattern = ''
+
+    potential_paths = [glob.glob(ec_path) for ec_path in create_paths(path, name, '*')]
+    potential_paths = sum(potential_paths, [])  # flatten
+
+    parsed_version = LooseVersion(version).version
+    version_patterns = [version]  # exact version match
+    if len(parsed_version) >= 2:
+        version_patterns.append(r'%s\.%s\.[0-9_A-Za-z]+' % tuple(parsed_version[:2]))  # major/minor version match
+    if parsed_version != parsed_version[0]:
+        version_patterns.append(r'%s\.[0-9-]+\.[0-9_A-Za-z]+' % parsed_version[0])  # major version match
+    version_patterns.append(r'[0-9._A-Za-z-]+')  # any version
+
+    regexes = []
+    for version_pattern in version_patterns:
+        regexes.extend([
+            re.compile((r"^\S+/%s-%s%s%s.eb$" % (name, version_pattern, toolchain_pattern, versionsuffix))),
+            re.compile((r"^\S+/%s-%s%s%s.eb$" % (name, version_pattern, toolchain_name_pattern, versionsuffix))),
+            re.compile((r"^\S+/%s-%s-\S+%s.eb$" % (name, version_pattern, versionsuffix))),
+            re.compile((r"^\S+/%s-%s%s.eb$" % (name, version_pattern, toolchain_pattern))),
+            re.compile((r"^\S+/%s-%s%s.eb$" % (name, version_pattern, toolchain_name_pattern))),
+            re.compile((r"^\S+/%s-%s-\S+.eb$" % (name, version_pattern))),
+        ])
+    _log.debug("found these potential paths: %s" % potential_paths)
+
+    for regex in regexes:
+        res = [p for p in potential_paths if regex.match(p)]
+        if res:
+            _log.debug("Related easyconfigs found using '%s': %s" % (regex.pattern, res))
+            break
+        else:
+            _log.debug("No related easyconfigs in potential paths using '%s'" % regex.pattern)
+    return res
+
+
 def review_pr(pull_request, colored=True, tmpdir=None):
     """Print multi-diff overview between easyconfigs in specified PR and current develop branch."""
     if tmpdir is None:
@@ -360,5 +435,9 @@ def review_pr(pull_request, colored=True, tmpdir=None):
     for ec in ecs:
         files = find_related_easyconfigs(repo_path, ec['ec'])
         _log.debug("File in pull request %s has these related easyconfigs: %s" % (ec['spec'], files))
-        diff = multi_diff(ec['spec'], files, colored=colored)
-        print diff
+        if files:
+            diff = multi_diff(ec['spec'], files, colored=colored)
+            msg = diff
+        else:
+            msg = "\n(no related easyconfigs found for %s)\n" % os.path.basename(ec['spec'])
+        print msg
